@@ -4,20 +4,20 @@
   "main": "main.py",
   "license": "GPL-3.0-only",
   "devDependencies": {
-    "bootstrap": "^5.3.2",
+    "bootstrap": "^5.3.3",
     "bootstrap-icons": "^1.11.3",
     "browserify": "^17.0.0",
-    "del": "^6.1.1",
-    "gulp": "^4.0.2",
+    "del": "^7.1.0",
+    "gulp": "^5.0.0",
     "gulp-rename": "^2.0.0",
     "gulp-replace": "^1.1.4",
     "gulp-sass": "^5.1.0",
     "gulp-uglify": "^3.0.2",
     "jquery": "^3.7.1",
-    "sass": "^1.69.7",
+    "sass": "^1.74.1",
     "svg-country-flags": "^1.2.10",
     "tsify": "^5.0.4",
-    "typescript": "^5.3.3",
+    "typescript": "^5.4.5",
     "vinyl-source-stream": "^2.0.0"
   }
 }
